tlwanstatus/requirements.txt
python-dotenv
selenium
bs4
click
